lab8/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.36
databases==0.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
starlette==0.41.0
httpx==0.27.2